# perf: Don't deepcopy Components, Tools, or Toolsets