--- a/haystack/components/agents/agent.py
+++ b/haystack/components/agents/agent.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import inspect
-from copy import deepcopy
 from typing import Any, Dict, List, Optional, Union
 
 from haystack import component, default_from_dict, default_to_dict, logging, tracing
@@ -11,6 +10,7 @@ from haystack.components.generators.chat.types import ChatGenerator
 from haystack.components.tools import ToolInvoker
 from haystack.core.pipeline.async_pipeline import AsyncPipeline
 from haystack.core.pipeline.pipeline import Pipeline
+from haystack.core.pipeline.utils import deepcopy_with_fallback
 from haystack.core.serialization import component_to_dict
 from haystack.dataclasses import ChatMessage
 from haystack.dataclasses.state import State, _schema_from_dict, _schema_to_dict, _validate_schema
@@ -111,7 +111,7 @@ class Agent:
         self._state_schema = state_schema or {}
 
         # Initialize state schema
-        resolved_state_schema = deepcopy(self._state_schema)
+        resolved_state_schema = deepcopy_with_fallback(self._state_schema)
         if resolved_state_schema.get("messages") is None:
             resolved_state_schema["messages"] = {"type": List[ChatMessage], "handler": merge_lists}
         self.state_schema = resolved_state_schema
@@ -244,7 +244,7 @@ class Agent:
             init_callback=self.streaming_callback, runtime_callback=streaming_callback, requires_async=False
         )
 
-        input_data = deepcopy({"messages": messages, "streaming_callback": streaming_callback, **kwargs})
+        input_data = deepcopy_with_fallback({"messages": messages, "streaming_callback": streaming_callback, **kwargs})
 
         state = State(schema=self.state_schema, data=kwargs)
         state.set("messages", messages)
@@ -332,7 +332,7 @@ class Agent:
             init_callback=self.streaming_callback, runtime_callback=streaming_callback, requires_async=True
         )
 
-        input_data = deepcopy({"messages": messages, "streaming_callback": streaming_callback, **kwargs})
+        input_data = deepcopy_with_fallback({"messages": messages, "streaming_callback": streaming_callback, **kwargs})
 
         state = State(schema=self.state_schema, data=kwargs)
         state.set("messages", messages)