--- a/haystack/components/agents/agent.py
+++ b/haystack/components/agents/agent.py
@@ -10,7 +10,7 @@ from haystack.components.generators.chat.types import ChatGenerator
 from haystack.components.tools import ToolInvoker
 from haystack.core.pipeline.async_pipeline import AsyncPipeline
 from haystack.core.pipeline.pipeline import Pipeline
-from haystack.core.pipeline.utils import deepcopy_with_fallback
+from haystack.core.pipeline.utils import _deepcopy_with_exceptions
 from haystack.core.serialization import component_to_dict
 from haystack.dataclasses import ChatMessage
 from haystack.dataclasses.state import State, _schema_from_dict, _schema_to_dict, _validate_schema
@@ -111,7 +111,7 @@ class Agent:
         self._state_schema = state_schema or {}
 
         # Initialize state schema
-        resolved_state_schema = deepcopy_with_fallback(self._state_schema)
+        resolved_state_schema = _deepcopy_with_exceptions(self._state_schema)
         if resolved_state_schema.get("messages") is None:
             resolved_state_schema["messages"] = {"type": List[ChatMessage], "handler": merge_lists}
         self.state_schema = resolved_state_schema
@@ -253,7 +253,7 @@ class Agent:
         with self._create_agent_span() as span:
             span.set_content_tag(
                 "haystack.agent.input",
-                deepcopy_with_fallback({"messages": messages, "streaming_callback": streaming_callback, **kwargs}),
+                _deepcopy_with_exceptions({"messages": messages, "streaming_callback": streaming_callback, **kwargs}),
             )
             counter = 0
             while counter < self.max_agent_steps:
@@ -342,7 +342,7 @@ class Agent:
         with self._create_agent_span() as span:
             span.set_content_tag(
                 "haystack.agent.input",
-                deepcopy_with_fallback({"messages": messages, "streaming_callback": streaming_callback, **kwargs}),
+                _deepcopy_with_exceptions({"messages": messages, "streaming_callback": streaming_callback, **kwargs}),
             )
             counter = 0
             while counter < self.max_agent_steps: