--- a/haystack/core/pipeline/utils.py
+++ b/haystack/core/pipeline/utils.py
@@ -14,46 +14,40 @@ from haystack.tools import Tool, Toolset
 logger = logging.getLogger(__name__)
 
 
-def deepcopy_with_fallback(obj: Any, max_depth: Optional[int] = 100) -> Any:
+def deepcopy_with_fallback(obj: Any) -> Any:
     """
-    Attempts to create a deep copy of the given object with a safe fallback mechanism.
+    Attempts to perform a deep copy of the given object.
 
-    If the object is a dictionary, each value is copied individually using this function recursively.
-    If an individual item fails to be copied, the original value is used as a fallback.
-    A maximum recursion depth is enforced to avoid deep or cyclic structures from causing issues.
+    This function recursively handles common container types (lists, tuples, sets, and dicts) to ensure deep copies
+    of nested structures. For specific object types that are known to be problematic for deepcopying—such as
+    instances of `Component`, `Tool`, or `Toolset` - the original object is returned as-is.
+    If `deepcopy` fails for any other reason, the original object is returned and a log message is recorded.
 
-    :param obj: The object to attempt to deep copy.
-    :param max_depth: The maximum depth to recurse during deep copying. If 0, returns the object as-is.
-    :return: A deep copy of the object if successful; otherwise, the original object.
+    :param obj: The object to be deep-copied.
+
+    :returns:
+        A deep-copied version of the object, or the original object if deepcopying fails.
     """
+    if isinstance(obj, list):
+        return [deepcopy_with_fallback(v) for v in obj]
+
+    if isinstance(obj, tuple):
+        return tuple(deepcopy_with_fallback(v) for v in obj)
+
+    if isinstance(obj, set):
+        return {deepcopy_with_fallback(v) for v in obj}
+
+    if isinstance(obj, dict):
+        return {k: deepcopy_with_fallback(v) for k, v in obj.items()}
+
     # Components and Tools often contain objects that we do not want to deepcopy or are not deepcopyable
     # (e.g. pytorch models, clients, etc.). In this case we return the object as-is.
     if isinstance(obj, (Component, Tool, Toolset)):
         return obj
 
-    if max_depth is not None and max_depth <= 0:
-        return obj
-
     try:
         return deepcopy(obj)
     except Exception as e:
-        # If the deepcopy fails we try to deepcopy each individual value if the object is a dictionary
-        next_depth = None if max_depth is None else max_depth - 1
-        if isinstance(obj, dict):
-            logger.info(
-                "Deepcopy failed for object of type '{obj_type}'. Error: {error}. Attempting item-wise copy.",
-                obj_type=type(obj).__name__,
-                error=e,
-            )
-            return {key: deepcopy_with_fallback(value, next_depth) for key, value in obj.items()}
-        elif isinstance(obj, (list, tuple, set)):
-            logger.info(
-                "Deepcopy failed for object of type '{obj_type}'. Error: {error}. Attempting item-wise copy.",
-                obj_type=type(obj).__name__,
-                error=e,
-            )
-            return type(obj)(deepcopy_with_fallback(item, next_depth) for item in obj)
-
         logger.info(
             "Deepcopy failed for object of type '{obj_type}'. Error: {error}. Returning original object instead.",
             obj_type=type(obj).__name__,