--- a/haystack/core/pipeline/utils.py
+++ b/haystack/core/pipeline/utils.py
@@ -5,7 +5,7 @@
 import heapq
 from copy import deepcopy
 from itertools import count
-from typing import Any, List, Optional, Tuple
+from typing import Any, List, Optional, Tuple, TypeVar
 
 from haystack import logging
 from haystack.core.component import Component
@@ -13,13 +13,15 @@ from haystack.tools import Tool, Toolset
 
 logger = logging.getLogger(__name__)
 
+T = TypeVar("T")
 
-def deepcopy_with_fallback(obj: Any) -> Any:
+
+def _deepcopy_with_exceptions(obj: T) -> T:
     """
     Attempts to perform a deep copy of the given object.
 
     This function recursively handles common container types (lists, tuples, sets, and dicts) to ensure deep copies
-    of nested structures. For specific object types that are known to be problematic for deepcopying—such as
+    of nested structures. For specific object types that are known to be problematic for deepcopying-such as
     instances of `Component`, `Tool`, or `Toolset` - the original object is returned as-is.
     If `deepcopy` fails for any other reason, the original object is returned and a log message is recorded.
 
@@ -29,10 +31,10 @@ def deepcopy_with_fallback(obj: Any) -> Any:
         A deep-copied version of the object, or the original object if deepcopying fails.
     """
     if isinstance(obj, (list, tuple, set)):
-        return type(obj)(deepcopy_with_fallback(v) for v in obj)
+        return type(obj)(_deepcopy_with_exceptions(v) for v in obj)
 
     if isinstance(obj, dict):
-        return {k: deepcopy_with_fallback(v) for k, v in obj.items()}
+        return {k: _deepcopy_with_exceptions(v) for k, v in obj.items()}
 
     # Components and Tools often contain objects that we do not want to deepcopy or are not deepcopyable
     # (e.g. models, clients, etc.). In this case we return the object as-is.