--- a/test/core/pipeline/test_utils.py
+++ b/test/core/pipeline/test_utils.py
@@ -251,22 +251,27 @@ class TestDeepcopyWithFallback:
         copy["objects"][1].name = "copied_obj2"
         assert copy["objects"][1].name == original["objects"][1].name
 
-    def test_deepcopy_with_fallback_tool(self):
+    def test_deepcopy_with_fallback_tool(self, caplog):
         tool = ComponentTool(
             name="problematic_tool", description="This is a problematic tool.", component=PromptBuilder("{{query}}")
         )
         original = {"tools": tool}
-        copy = deepcopy_with_fallback(original)
+        with caplog.at_level(logging.INFO):
+            copy = deepcopy_with_fallback(original)
+            assert "Deepcopy failed for object of type 'dict'" in caplog.text
         assert copy["tools"] == original["tools"]
         # Not a copy so changing the name in the copy should also affect the original
         copy["tools"].name = "copied_tool"
         assert copy["tools"] == original["tools"]
 
-    def test_deepcopy_with_fallback_component(self, monkeypatch):
+    def test_deepcopy_with_fallback_component(self, monkeypatch, caplog):
         monkeypatch.setenv("OPENAI_API_KEY", "test")
         comp = OpenAIChatGenerator()
-        res = deepcopy_with_fallback(comp)
-        assert res == comp
+        original = {"component": comp}
+        with caplog.at_level(logging.INFO):
+            res = deepcopy_with_fallback(original)
+            assert "Deepcopy failed for object of type 'dict'" in caplog.text
+        assert res == original
         # Change an attribute of the copy and check original also updates
-        res.model = "a-model"
-        assert res.model == comp.model
+        res["component"].model = "a-model"
+        assert res["component"].model == original["component"].model