### Related Issues

- partially address https://github.com/deepset-ai/haystack/issues/9349

### Proposed Changes:

 <!--- In case of a bug: Describe what caused the issue and how you solved it -->
 <!--- In case of a feature: Describe what did you add and how it works -->
In our pipeline run methods we deepcopy the inputs and outputs of components to ensure that we don't cause unintended behavior such as altering the Pipeline run logic unepxectedly when the same input reference is passed to multiple components. 

This has worked well under the assumption that all run time inputs and outputs are data-like objects (e.g. strings, dataclasses, numbers, etc.) but has recently caused issues with the introduction of being able to pass Tools as runtime parameters to ChatGenerators. Since Tools can and often contain Components (e.g. ComponentTool) or even full pipelines (e.g. SuperComponent wrapped by ComponentTool) we find that our Pipeline run logic tries to deepcopy said tool leading to deepcopying of Components, Document Stores, etc. 

This has unsurprisingly caused issues since many of our Components contain attributes that aren't deepcopyable such as Jinja2 templates or some clients (e.g. ProxyClient). Or they contain objects that'd we'd prefer not to deepcopy such as local ML models.  

So I think it makes sense to update the new utility method `deepcopy_with_fallback` to automatically skip over Component, Tool or Toolset if they are passed into the deepcopy. 

### How did you test it?

<!-- unit tests, integration tests, manual verification, instructions for manual tests -->
- Updated unit tests.
- Also manually triggered end to end tests https://github.com/deepset-ai/haystack/actions/runs/14905702825

### Notes for the reviewer

<!-- E.g. point out section where the reviewer  -->

Some performance checks on the deepcopying
```python
import time
import tracemalloc

from haystack.core.pipeline.utils import deepcopy_with_fallback
from haystack.components.rankers.transformers_similarity import TransformersSimilarityRanker
from haystack.tools import ComponentTool
from haystack.utils import ComponentDevice


st = TransformersSimilarityRanker(device=ComponentDevice.from_str("cpu"))
st.warm_up()

tracemalloc.start()
start_time = time.time()

comp_tool = ComponentTool(
    component=st,
    name="text_embedder",
    description="Use this tool to embed text.",
)

_ = deepcopy_with_fallback(
    {
        "tools": [
            comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool
        ]
    }
)

end_time = time.time()
current, peak = tracemalloc.get_traced_memory()
tracemalloc.stop()

print(f"Time elapsed: {end_time - start_time:.4f} seconds")
print(f"Current memory usage: {current / 1024**2:.2f} MB; Peak: {peak / 1024**2:.2f} MB")

# When skipping copying of Tool 
# Ten tools
# Time elapsed: 0.0273 seconds
# Current memory usage: 0.14 MB; Peak: 0.16 MB

# When deepcopying Tool
# Ten tools
# Time elapsed: 0.4356 seconds
# Current memory usage: 3.81 MB; Peak: 4.54 MB
```

Another example with a custom component that contains an attribute that causes a circular deepcopy
```python
import time
import tracemalloc

import logging
logging.basicConfig(level=logging.INFO)

from haystack import component
from haystack.core.pipeline.utils import deepcopy_with_fallback
from haystack.tools import ComponentTool
from ddtrace.settings._config import Config


@component
class BadComponent:
    def __init__(self):
        # Object that can't be deepcopied
        self.config = Config()

    @component.output_types(result=str)
    def run(self, inp: str):
        return inp


bad_component = BadComponent()

tracemalloc.start()
start_time = time.time()

comp_tool = ComponentTool(
    component=bad_component,
    name="bad_component",
    description="Use this tool to return text.",
)

# _ = deepcopy_with_fallback({"tools": [comp_tool]})
_ = deepcopy_with_fallback(
    {
        "tools": [
            comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool, comp_tool
        ]
    }
)

end_time = time.time()
current, peak = tracemalloc.get_traced_memory()
tracemalloc.stop()

print(f"Time elapsed: {end_time - start_time:.4f} seconds")
print(f"Current memory usage: {current / 1024**2:.2f} MB; Peak: {peak / 1024**2:.2f} MB")

# When using new version
# One tool
# Time elapsed: 0.0033 seconds
# Current memory usage: 0.06 MB; Peak: 0.06 MB
# Ten tools
# Time elapsed: 0.0036 seconds
# Current memory usage: 0.06 MB; Peak: 0.06 MB

# When using old version
# One tool
# Time elapsed: 0.0611 seconds
# Current memory usage: 0.15 MB; Peak: 0.56 MB
# Ten tools
# Time elapsed: 0.6274 seconds
# Current memory usage: 0.15 MB; Peak: 0.57 MB
```

### Checklist

- I have read the [contributors guidelines](https://github.com/deepset-ai/haystack/blob/main/CONTRIBUTING.md) and the [code of conduct](https://github.com/deepset-ai/haystack/blob/main/code_of_conduct.txt)
- I have updated the related issue with new insights and changes
- I added unit tests and updated the docstrings
- I've used one of the [conventional commit types](https://www.conventionalcommits.org/en/v1.0.0/) for my PR title: `fix:`, `feat:`, `build:`, `chore:`, `ci:`, `docs:`, `style:`, `refactor:`, `perf:`, `test:` and added `!` in case the PR includes breaking changes.
- I documented my code
- I ran [pre-commit hooks](https://github.com/deepset-ai/haystack/blob/main/CONTRIBUTING.md#installation) and fixed any issue