@sjrl 2025-05-08
> Is it difficult to change this aspect?

Hmm it might not be if we decide to only check for a few base types to iterate through like dict, list, tuple, and set.