@sjrl 2025-05-08
> I have nothing against this idea, if it does not break stuff.
> 
> In general, since we are talking about performance, it would be nice to see some (even simple) benchmarks: maybe running one/a few Pipelines with `ComponentTool` before and after this PR.
> 
> > ... the way the function is currently set up it will first try to deepcopy inputs and then only if it fails with an exception it will iterate through each key-value pair in this case and try to deepcopy the values. Only then would it skip over each individual tool.
> 
> Is it difficult to change this aspect?

@anakin87 I was able to refactor this to solve this issue