@sjrl 2025-05-08
And I agree on adding performance benchmarks. This isn't urgent since the original version of `deepcopy_with_fallback` should already unblock errors we've been seeing, just wanted to get this out while it was fresh in my head.