@sjrl 2025-05-08
@anakin87 I've added some performance examples.

My interpretation of the results is that the performance impact linearly scales with the number of problematic components provided through tools. My examples are relatively simple but show how going from one problematic tool to ten can cause a noticeable delay. In this case the delay would happen every time the pipeline is run. 

And this problem only gets worse and worse if users provide SuperComponents as Tools or Agents as Tools and we get a continuing nested nature of components getting copied.